Updated docs with new methods
